Fix four simulation bugs + repair the main build (#123)
* Fix the main build: three compile errors on origin/main

main does not compile, and has not since PR #97. Every EditMode run and every
player build fails. Cloud agents cannot run Unity, which is the gap AGENTS.md
already warns about -- a "tests pass" line from one of those PRs is unverified by
construction. Carried here because the correctness fixes on this branch cannot be
verified until main compiles again.

- AirsidePrefabAddressables implements IResourceLocator but imported only
  UnityEngine.AddressableAssets; the interface is in
  UnityEngine.AddressableAssets.ResourceLocators (CS0246).

- Its Locator did not implement IResourceLocator.AllLocations, which this
  Addressables version requires under the binary catalog (CS0535). Implemented,
  guarded by the same !ENABLE_JSON_CATALOG the interface uses.

- It also referenced LegacyResourcesProvider, which this Addressables version does
  not ship at all (CS0246). Since the file has never compiled, this locator has
  never run: ArtPresentationLoader's Resources -> glTF fallbacks are what have
  actually been serving prefabs. The provider is now named rather than resolved via
  typeof, and EnsureRegistered is gated behind a new Enabled flag that is off --
  registering a locator whose provider cannot load would turn a working fallback
  into a runtime failure. Bailey's Editor Addressables groups are the intended
  replacement; flip Enabled when a real provider is wired.

- AirsidePrototype passed AirportRoutes.FirstOfferAfterSeconds (long) into the
  HUD overlays' int parameter (CS1503, twice).

Unity EditMode after this commit: green.


* Fix four simulation bugs that contradicted the design brief

Each fix ships with a regression test verified failing against the previous code.

- Ground traffic taxied to and parked on Stand 3 before the player built it.
  AlternateStand walked Stand 1/2/3 with no knowledge of Capacity.StandCount, so
  once two commercials filled the baseline stands GT-201 reserved LEAD-IN-3 and
  parked on bare ground. Stand choice is now capacity-aware, and GT-201 holds
  off-field -- leaving the corridor free for GT-202 -- when every built stand is
  occupied.

- Understaffing delays were reported with no cause. The cause only ever checked
  HasCleaningDisruption, so a short-crewed turnaround logged "Delayed 12s" with an
  empty detail, against the brief's "every delay should have an understandable
  cause". TurnaroundWorkflow.OverrunCause now names cleaning disruption,
  understaffing, or a generic extended turnaround.

- The daily finance brief halved once a second commercial started operating: the
  projection assumed exactly one aircraft flying the cycle, so it under-reported
  income and the cash runway by 2x precisely when the player had scaled up.

- A player command issued at simulation second 0 was dropped on reload. ReplayTo
  advances the clock then applies commands at the new second, so it never visited
  second 0 -- the cash it cost was replayed but the effect was not.

Unity EditMode on the Mac editor: 116/116.


---------

+++ game/Airside/Assets/Airside/Presentation/AirsidePrototype.cs
@@ -399,7 +399,7 @@ namespace Airside.Presentation
                     showAway: showAway,
                     showInsolvency: showInsolvency,
                     locationName: _simulation.Location.Name,
-                    firstOfferAfterSeconds: AirportRoutes.FirstOfferAfterSeconds,
+                    firstOfferAfterSeconds: (int)AirportRoutes.FirstOfferAfterSeconds,
                     awayBody: awayBody,
                     insolvencyBody: insolvencyBody);
                 _canvasHud.SetOverlaysVisible(false);
@@ -414,7 +414,7 @@ namespace Airside.Presentation
                     showAway: showAway,
                     showInsolvency: showInsolvency,
                     locationName: _simulation.Location.Name,
-                    firstOfferAfterSeconds: AirportRoutes.FirstOfferAfterSeconds,
+                    firstOfferAfterSeconds: (int)AirportRoutes.FirstOfferAfterSeconds,
                     awayBody: awayBody,
                     insolvencyBody: insolvencyBody);
             }
